+++ frontend/src/app/policy-engine/export-import-dialog/export-import-dialog.component.css
@@ -1,0 +1,74 @@
+form {
+    display: flex;
+    flex-direction: column;
+    width: 452px;
+    overflow: visible;
+    padding-bottom: 25px;
+}
+
+.context {
+    position: relative;
+    overflow-y: auto;
+    padding: 50px 70px 50px 70px;
+}
+
+.loading {
+    background: #fff;
+    position: absolute;
+    z-index: 999;
+    left: 0;
+    bottom: 0;
+    right: 0;
+    top: 0;
+    display: flex;
+    align-items: center;
+    justify-items: center;
+    justify-content: center;
+    align-content: center;
+}
+
+.policy-header {
+    height: 44px;
+    font-size: 34px;
+    display: block;
+    padding: 12px;
+    box-sizing: border-box;
+    text-align: center;
+    color: #000;
+}
+
+.delimiter {
+    padding: 55px 55px;
+    width: 100%;
+    height: 110px;
+    box-sizing: border-box;
+    position: relative;
+}
+
+.delimiter::after {
+    content: '';
+    position: absolute;
+    height: 2px;
+    top: 55px;
+    left: 0px;
+    right: 0px;
+    border-top: 1px solid #c8c8c8;
+}
+
+.field-value {  
+    color: #000;
+    font-size: 18px;
+}
+
+.field-name {  
+    margin-bottom: 8px;
+}
+
+.field {
+    margin-top: 5px;
+    margin-bottom: 28px;
+}
+
+.field-last {
+    margin-bottom: 0px;
+}
