--- frontend/src/app/policy-engine/export-policy-dialog/export-policy-dialog.component.css
@@ -1,79 +1,0 @@
-form {
-    display: flex;
-    flex-direction: column;
-    width: 452px;
-    overflow: visible;
-    padding-bottom: 25px;
-}
-
-.context {
-    position: relative;
-    height: 500px;
-    overflow-y: auto;
-    padding-bottom: 10px;
-}
-
-.loading {
-    background: #fff;
-    position: absolute;
-    z-index: 999;
-    left: 0;
-    bottom: 0;
-    right: 0;
-    top: 0;
-    display: flex;
-    align-items: center;
-    justify-items: center;
-    justify-content: center;
-    align-content: center;
-}
-
-.dialog-actions {
-    padding: 8px 0px;
-}
-
-.dialog-actions button {
-    margin-right: 8px;
-}
-
-.field {
-    display: block;
-}
-.field-name {
-    display: inline-block;
-    margin-right: 8px;
-}
-.field-value {
-    display: inline-block;
-    margin-right: 8px;
-}
-
-.field-header {
-    font-size: 16px;
-    font-weight: 500;
-    padding: 8px 0px;
-}
-
-.schema-item {
-    padding: 4px 14px;
-}
-
-.field-relationships {
-    display: block;
-    padding-left: 25px;
-    font-size: 12px;
-    color: #0C77FF;
-}
-
-.token-item {
-    padding: 4px 14px;
-}
-    
-.field-keys {
-    display: block;
-    padding-left: 25px;
-}
-
-.policy .field {
-    padding: 2px 14px;
-}
